gcf_video_summary/gcf2_generate_summary/requirements.txt
# https://pypi.org/project/google-cloud-storage/
google-cloud-storage==1.42.3

# https://pypi.org/project/opencv-python-headless/
opencv-python-headless==4.5.3.56

# https://pypi.org/project/Pillow/
Pillow==8.3.2
